WIP - abstract form-base.components.ts with search-form.component.ts only
