+++ src/app/components/index.ts
@@ -26,6 +26,7 @@ import { GovukBackLinkComponent } from './govuk/govuk-back-link/govuk-back-link.
 import { GovukTagComponent } from './govuk/govuk-tag/govuk-tag.component';
 import { GovukTaskListComponent } from './govuk/govuk-task-list/govuk-task-list.component';
 import { GovukTaskListItemComponent } from './govuk/govuk-task-list/govuk-task-list-item/govuk-task-list-item.component';
+import { FormBaseComponent } from './abstract/form-base/form-base.component';
 
 export {
   GovukButtonComponent,
@@ -56,4 +57,5 @@ export {
   GovukTagComponent,
   GovukTaskListComponent,
   GovukTaskListItemComponent,
+  FormBaseComponent,
 };
