+++ src/app/mocks/index.ts
@@ -22,6 +22,7 @@ import { AUTO_COMPLETE_ITEMS_MOCK } from './auto-complete-items.mock';
 import { FORM_CONTROL_ERROR_MOCK } from './form-control-error.mock';
 import { FORM_ERROR_SUMMARY_MOCK } from './form-error-summary-entry.mock';
 import { FORM_DATE_ERROR_SUMMARY_MOCK } from './form-date-error-summary.mock';
+import { FIELD_ERROR_MOCK } from './field-errors.mock';
 
 export {
   DEFENDANT_ACCOUNT_MOCK,
@@ -48,4 +49,5 @@ export {
   FORM_CONTROL_ERROR_MOCK,
   FORM_ERROR_SUMMARY_MOCK,
   FORM_DATE_ERROR_SUMMARY_MOCK,
+  FIELD_ERROR_MOCK,
 };
