+++ src/app/pages/account-enquiry/search/search-form/search-form.component.html
@@ -7,9 +7,9 @@
       [errors]="formErrorSummaryMessage"
       (errorClick)="scrollTo($event)"
     ></app-govuk-error-summary>
-    <div [formGroup]="searchForm">
+    <div [formGroup]="form">
       <app-alphagov-accessible-autocomplete
-        [control]="searchForm.controls['court']"
+        [control]="form.controls['court']"
         labelText="Court"
         inputId="court"
         inputName="court"
@@ -24,24 +24,24 @@
         inputId="surname"
         inputName="surname"
         inputClasses="govuk-!-width-one-half"
-        [control]="searchForm.controls['surname']"
+        [control]="form.controls['surname']"
       ></app-govuk-text-input>
       <app-govuk-text-input
         labelText="Forename"
         inputId="forename"
         inputName="forename"
         inputClasses="govuk-!-width-one-half"
-        [control]="searchForm.controls['forename']"
+        [control]="form.controls['forename']"
       ></app-govuk-text-input>
       <app-govuk-text-input
         labelText="Initials"
         inputId="initials"
         inputName="initials"
         inputClasses="govuk-!-width-one-quarter"
-        [control]="searchForm.controls['initials']"
+        [control]="form.controls['initials']"
       ></app-govuk-text-input>
       <app-govuk-date-input
-        [group]="searchForm.get('dateOfBirth')"
+        [group]="form.get('dateOfBirth')"
         [dateInputs]="dateInputs"
         fieldSetId="dateOfBirth"
         legendText="Date of Birth"
@@ -55,21 +55,21 @@
         inputId="addressLine"
         inputName="addressLine"
         inputClasses="govuk-!-width-one-half"
-        [control]="searchForm.controls['addressLine']"
+        [control]="form.controls['addressLine']"
       ></app-govuk-text-input>
       <app-govuk-text-input
         labelText="NI Number"
         inputId="niNumber"
         inputName="niNumber"
         inputClasses="govuk-!-width-one-half"
-        [control]="searchForm.controls['niNumber']"
+        [control]="form.controls['niNumber']"
       ></app-govuk-text-input>
       <app-govuk-text-input
         labelText="PCR"
         inputId="pcr"
         inputName="pcr"
         inputClasses="govuk-!-width-one-half"
-        [control]="searchForm.controls['pcr']"
+        [control]="form.controls['pcr']"
       ></app-govuk-text-input>
     </div>
   </fieldset>
